+++ app/src/main/java/com/coinhub/android/data/models/TopUpModel.kt
@@ -1,6 +1,6 @@
 package com.coinhub.android.data.models
 
-import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.LocalDate
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
@@ -8,7 +8,7 @@ import kotlin.uuid.Uuid
 data class TopUpModel @OptIn(ExperimentalUuidApi::class) constructor(
     val id: Uuid,
     val provider: TopUpProviderEnum,
-    val amount: BigDecimal,
+    val amount: BigInteger,
     val status: TopUpStatusEnum,
     val createdAt: LocalDate,
 )
